// src/types/movieType.ts
export interface Movie {
    id: string;
    title: string;
    original_title: string;
    original_title_romanized: string;
    image: string;
    movie_banner: string;
    description: string;
    director: string;
    producer: string;
    release_date: string;
    running_time: string;
    rt_score: string;
    people: string[];
    species: string[];
    locations: string[];
    vehicles: string[];
    url: string;
}